requirements.txt
streamlit
pillow
numpy
joblib
opencv-python